Last sequence is written
Added the writeLast method to bitPacker. If there is something left in the buffer, but we are done with packing, we form the byte out of buffer and put in into file.  In banana_bandana case everything works fine. In other with two last sequnces being identical still loosing the last one