Special cases for Decoder ~
Added the special cases when we read the same sequence we just added and so on. Changed the part in Encoder, when we are writting the last sequence, now writing the prevStr, so the whole last sequnce, seems to work. In dictionary added the getNextInd method, but probably it has wrong logic and shoud not be with +1. TODO: improve the special cases handle and go to bitReader